+++ test-src/test.js
@@ -1,4 +1,4 @@
-import {module} from "../mock-injects"
+import {module} from "../src/mock-injects"
 import testOfflineStorage from "./test-offline-storage"
 
 const testApp = angular.module('test-ack-angular',['ngMockE2E']).name