+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # ack-angular - Change Log
 All notable changes to this project will be documented here.
 
+## [1.7.0] - 2016-12-12
+### Enhanced
+- angular 1.6 requires more use of $onInit for setting initial model values
+### Breaking change
+- switch JSPM for WEBPACK
+
 ## [1.6.2] - 2016-10-07
 ### Added
 - ack-state.module